Test response
Added if/elif statements to the /process route. This is to see if while guessing the answer, will the correct response come up when submitted. After changing session['guess'] in an int type from a unicode type, no issues. Will change to variable to display on index.html rather than just the terminal